requirements.txt
# PM-JIA Requirements
openai==1.97.1
google-genai==1.29.0
tavily-python==0.7.10
fastapi==0.115.9
uvicorn==0.34.3
click==8.2.1
rich==14.0.0
pandas==2.2.3
pydantic==2.11.5
python-dotenv==1.1.0
PyYAML==6.0.2
Pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0